project/requirements.txt
Django==5.0
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.14.0
djoser==2.2.2
social-auth-app-django==5.4.0
django-ckeditor==6.7.0
pytz==2023.3
Pillow==10.1.0
mutagen==1.47.0
psycopg2==2.9.9
argon2-cffi==21.3.0 
braintree==4.24.0
pyJWT==2.1.0
boto3==1.33.13
django-storages==1.14.2 
markdown==3.5.1
gunicorn==21.2.0
whitenoise==6.6.0
requests==2.31.0
python-slugify==8.0.1


